Record max head memory
Fixes #731
